Ai-question-recommendation/requirements.txt
streamlit
google-generativeai
pdfminer.six
docx2txt
